+++ shinobi_client/tests/orms/test_user.py
@@ -1,9 +1,8 @@
 import unittest
 
-from shinobi_client.api_key import ShinobiApiKey, ShinobiWrongPasswordError
-from shinobi_client.orms.user import ShinobiUserOrm
+from shinobi_client.api_key import ShinobiApiKey
+from shinobi_client.orms.user import ShinobiUserOrm, ShinobiWrongPasswordError
 from shinobi_client.tests._common import _create_email_and_password, TestWithShinobi
-from shinobi_client.client import ShinobiClient
 from shinobi_client._common import ShinobiSuperUserCredentialsRequiredError
 
 
@@ -14,6 +13,7 @@ class TestShinobiUserOrm(TestWithShinobi):
     def setUp(self):
         super().setUp()
         self.user_orm = ShinobiUserOrm(self._shinobi_client)
+        self.superless_user_orm = ShinobiUserOrm(self._superless_shinobi_client)
         self.api_key = ShinobiApiKey(self._shinobi_client)
 
     def test_create(self):
@@ -22,6 +22,10 @@ class TestShinobiUserOrm(TestWithShinobi):
         user = self.user_orm.get(email)
         self.assertEqual(email, user["mail"])
         self.assertIsNotNone(self.api_key.get(email, password))
+
+    def test_create_requires_super_user_credentials(self):
+        email, password = _create_email_and_password()
+        self.assertRaises(ShinobiSuperUserCredentialsRequiredError, self.superless_user_orm.create, email, password)
 
     def test_get_when_does_not_exist(self):
         user = self.user_orm.get("example@doesnotexist.com")
@@ -32,6 +36,11 @@ class TestShinobiUserOrm(TestWithShinobi):
         retrieved_user = self.user_orm.get(user["email"])
         self.assertEqual(user["mail"], retrieved_user["mail"])
 
+    def test_get_using_user_credentials(self):
+        user = self._create_user()
+        retrieved_user = self.superless_user_orm.get(user["email"], user["password"])
+        self.assertEqual(user["mail"], retrieved_user["mail"])
+
     def test_get_all(self):
         emails = []
         for i in range(3):
@@ -39,6 +48,9 @@ class TestShinobiUserOrm(TestWithShinobi):
 
         matched_users = tuple(filter(lambda user: user["mail"] in emails, self.user_orm.get_all()))
         self.assertEqual(len(emails), len(matched_users))
+
+    def test_get_all_requires_super_user_credentials(self):
+        self.assertRaises(ShinobiSuperUserCredentialsRequiredError, self.superless_user_orm.get_all)
 
     def test_modify_password_when_user_not_exist(self):
         email, password = _create_email_and_password()
@@ -61,11 +73,6 @@ class TestShinobiUserOrm(TestWithShinobi):
         user = self._create_user()
         self.user_orm.delete(user["email"])
         self.assertIsNone(self.user_orm.get(user["email"]))
-
-    def test_requires_super_user_credentials(self):
-        superless_shinobi_client = ShinobiClient(self._shinobi_client.host, self._shinobi_client.port)
-        superless_user_orm = ShinobiUserOrm(superless_shinobi_client)
-        self.assertRaises(ShinobiSuperUserCredentialsRequiredError, superless_user_orm.get_all)
 
 
 if __name__ == "__main__":